relay_conventions

Module measurements

Measurement constant definitions.

Constants

APP_START_COLD

The time it takes for the app to start from a cold state.

APP_START_WARM

The time it takes for the app to start from a warm state.

CLS

Tthe sum total of all individual layout shift scores for every unexpected layout shift that occurs during the entire lifespan of the page.

FCP

The time it takes for the browser to render the first piece of meaningful content on the screen.

FID

The time from when a user first interacts with a page to the time when the browser is actually able to begin processing event handlers in response to that interaction.

FP

The time it takes for the browser to render the first pixel on the screen.

FRAMES_FROZEN

The number of frames that took longer than 700ms to render.

FRAMES_FROZEN_RATE

measurements.frames_frozen divided by the measurements.frames_total.

FRAMES_SLOW

The number of frames that took longer than 16ms (Android) or 16.67 (iOS) to render.

FRAMES_SLOW_RATE

measurements.frames_slow divided by the measurements.frames_total.

FRAMES_TOTAL

Total frames rendered.

INP

The time it takes for the browser to render the next frame after a user interacts with the page.

LCP

The time it takes for the largest contentful paint element to be rendered on the screen.

STALL_COUNT

A stall is when the JavaScript event loop takes longer than expected to complete. Only applies to React Native.

STALL_LONGEST_TIME

The longest stall time is the time of the longest event loop stall. Only applies to React Native.

STALL_PERCENTAGE

Stall percentage is equal to the stall_total_time divided by the transaction.duration. Only applies to React Native.

STALL_TOTAL_TIME

The total stall time is the total combined time of all stalls. Only applies to React Native.

TTFB

The time it takes for the first byte of a response from the server to reach the user’s browser after a request is made.

TTFB__REQUESTTIME

The time it takes for the server to process the initial request and send the first byte of a response to the user’s browser.