Crate relay_conventions

Attribute definitions extracted from sentry-conventions.

This crate contains the sentry-conventions repository as a git submodule. Attribute definitions in the submodule are parsed at compile time and can be accessed via the attribute_info function.

It also exposes a number of constants for attribute names that Relay has specific logic for. It is recommended to use these constants instead of the bare attribute names to ensure consistency.

Structs

AttributeInfo

Information about an attribute, as defined in sentry-conventions.

Enums

Pii

Whether an attribute should be PII-strippable/should be subject to datascrubbers

WriteBehavior

Under which names an attribute should be saved.

Constants

BROWSER_NAME

BROWSER_VERSION

CLIENT_ADDRESS

CLIENT_SAMPLE_RATE

DB_QUERY_TEXT

DB_STATEMENT

DB_SYSTEM_NAME

DESCRIPTION

ENVIRONMENT

FAAS_TRIGGER

GEN_AI_SYSTEM

GRAPHQL_OPERATION

HTTP_PREFETCH

HTTP_REQUEST_METHOD

HTTP_RESPONSE_STATUS_CODE

HTTP_ROUTE

HTTP_TARGET

MESSAGING_SYSTEM

OBSERVED_TIMESTAMP_NANOS

OP

ORIGIN

PLATFORM

PROFILE_ID

RELEASE

RPC_GRPC_STATUS_CODE

RPC_SERVICE

SEGMENT_ID

SEGMENT_NAME

SERVER_ADDRESS

STATUS_MESSAGE

URL_FULL

URL_PATH

USER_AGENT_ORIGINAL

USER_GEO_CITY

USER_GEO_COUNTRY_CODE

USER_GEO_REGION

USER_GEO_SUBDIVISION

Functions

attribute_info

Returns information about an attribute, as defined in sentry-conventions.

name_for_op_and_attributes