Crate relay_conventions

Attribute definitions extracted from sentry-conventions.

This crate contains the sentry-conventions repository as a git submodule. Attribute definitions in the submodule are parsed at compile time and can be accessed via the attribute_info function.

It also exposes a number of constants for attribute names that Relay has specific logic for. It is recommended to use these constants instead of the bare attribute names to ensure consistency.

Re-exports

pub use consts::*;

Modules

consts

Structs

AttributeInfo

Information about an attribute, as defined in sentry-conventions.

Enums

Pii

Whether an attribute should be PII-strippable/should be subject to datascrubbers

WriteBehavior

Under which names an attribute should be saved.

Functions

attribute_info

Returns information about an attribute, as defined in sentry-conventions.

name_for_op_and_attributes